requirements.txt
Scrapy==1.8.0
sqlalchemy==1.3.13
psycopg2-binary==2.8.4
python-dotenv==0.11.0